% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_key}
\alias{cache_key}
\title{Generate Cache Key}
\usage{
cache_key(tag, X)
}
\arguments{
\item{tag}{Character string tag}

\item{X}{Data object to hash}
}
\value{
Cache key string
}
\description{
Generate Cache Key
}
\keyword{internal}
