% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cached}
\alias{get_cached}
\title{Retrieve Cached Object}
\usage{
get_cached(tag, X)
}
\arguments{
\item{tag}{Character string tag}

\item{X}{Data object (for key generation)}
}
\value{
Cached object or NULL if not found
}
\description{
Retrieve Cached Object
}
\keyword{internal}
