% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mse}
\alias{mse}
\title{Calculate Mean Squared Error}
\usage{
mse(a, b)
}
\arguments{
\item{a}{Numeric vector or convertible to numeric}

\item{b}{Numeric vector or convertible to numeric}
}
\value{
Mean squared error between a and b
}
\description{
Calculate Mean Squared Error
}
\keyword{internal}
