% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimv_diagnostics.R
\name{plot_feature_selection}
\alias{plot_feature_selection}
\title{Plot Feature Relevance Scores}
\usage{
plot_feature_selection(score_df)
}
\arguments{
\item{score_df}{A data frame output from \code{feature_select_score()}.}
}
\value{
A \code{ggplot} object for visualization.
}
\description{
Visualizes feature ranking generated by \code{feature_select_score()}.
This plot can help assess which variables are most influential for
DIMV conditional modeling.
}
