% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimv_diagnostics.R
\name{print_dimv_diag}
\alias{print_dimv_diag}
\title{Print DIMV Convergence Diagnostics}
\usage{
print_dimv_diag(imputer)
}
\arguments{
\item{imputer}{A \code{dimv_imputer} object.}
}
\description{
Pretty-print wrapper for \code{dimv_convergence_diag()} to provide a
clean console summary of model convergence and variance behaviour.
}
