% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_adaptive_threshold}
\alias{select_adaptive_threshold}
\title{Adaptive threshold selection (correlation-based)}
\usage{
select_adaptive_threshold(abs_cors, min_features = 2, max_features = 5)
}
\description{
Adaptive threshold selection (correlation-based)
}
\keyword{internal}
