% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_fixed_threshold}
\alias{select_fixed_threshold}
\title{Fixed threshold selection (correlation-based)}
\usage{
select_fixed_threshold(
  abs_cors,
  threshold = 0.3,
  min_features = 2,
  max_features = 5
)
}
\description{
Fixed threshold selection (correlation-based)
}
\keyword{internal}
