% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_hybrid}
\alias{select_hybrid}
\title{Hybrid selection (correlation + MI)}
\usage{
select_hybrid(X, target_var, min_features = 2, max_features = 5, nbins = 5)
}
\description{
Hybrid selection (correlation + MI)
}
\keyword{internal}
