% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_mutual_information}
\alias{select_mutual_information}
\title{Mutual information selection}
\usage{
select_mutual_information(
  X,
  target_var,
  min_features = 2,
  max_features = 5,
  nbins = 5
)
}
\description{
Mutual information selection
}
\keyword{internal}
