% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_cached}
\alias{set_cached}
\title{Store Object in Cache}
\usage{
set_cached(tag, X, obj)
}
\arguments{
\item{tag}{Character string tag}

\item{X}{Data object (for key generation)}

\item{obj}{Object to cache}
}
\value{
NULL (invisibly)
}
\description{
Store Object in Cache
}
\keyword{internal}
