% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{simulate_missing}
\alias{simulate_missing}
\title{Simulate Missing Data (MCAR/MAR/MNAR) safely}
\usage{
simulate_missing(X, rate = 0.2, mechanism = "MCAR")
}
\description{
Simulate Missing Data (MCAR/MAR/MNAR) safely
}
\keyword{internal}
