% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{within_var_mse}
\alias{within_var_mse}
\title{Compute Within-Imputation Variance for MSE of predictions}
\usage{
within_var_mse(y_true, y_hat)
}
\description{
Compute Within-Imputation Variance for MSE of predictions
}
\keyword{internal}
