% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{within_var_shap_mse}
\alias{within_var_shap_mse}
\title{Within-variance for SHAP MSE (elementwise over matrix)}
\usage{
within_var_shap_mse(shap_mat, ref_shap)
}
\description{
Within-variance for SHAP MSE (elementwise over matrix)
}
\keyword{internal}
